refactor(workflows): fold single-consumer services into workflow folders
doc-drift-tasks, prod-log-triage-tasks, pr-extract-tasks, reviews-inbox and
publish-review each had exactly one consumer: their workflows/<kind>/ folder.
Move them in (completing the per-kind consolidation), fix relative imports,
and update the stale publish-review path references in the review test mock,
playwright config and e2e helper.
