+++ server/workflows/doc-drift/doc-drift-tasks.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach } from 'vitest';
-import { createTestDb } from './test-helpers.js';
-import { getDb } from './db.js';
+import { createTestDb } from '../../test-helpers.js';
+import { getDb } from '../../db.js';
 import { buildDocDriftPrompt, insertDocDriftTask } from './doc-drift-tasks.js';
 
 describe('buildDocDriftPrompt', () => {
