+++ server/workflows/reviewer/reviews-inbox.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach } from 'vitest';
-import { createTestDb } from './test-helpers.js';
-import { getDb } from './db.js';
+import { createTestDb } from '../../test-helpers.js';
+import { getDb } from '../../db.js';
 import { listReviewsInbox, getReviewDetail } from './reviews-inbox.js';
 
 function seedRepo(): void {
